你把钱借给了你的朋友，那么他一定会记得你，在他下一次缺钱时。

+++ euler/euler407/main.c
@@ -37,9 +37,9 @@ uint32_t *findLargestPrimes(uint32_t limit) {
                 }
 
                 // and all exponents i^2, i^3, i^4 ...
-                uint32_t j = i * i;
+                uint64_t j = (uint64_t)i * i;
                 while (j <= limit) {
-                        vec[j] = j;
+                        vec[j] = (uint32_t)j;
                         j *= i;
                 }
         }
